color: prioritize protocol error instead of graceful failure
In the parametric builder API, which is used by our color management
protocol implementation, we have errors that may be translated to
protocol errors or graceful failures. As we return a single err code
to the protocol implementation, let's prioritize the protocol error.

We do that by inverting the order of checks, as the first error code
caught is the one prioritized.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>