timeline: Post core_flush_damage after clearing damage
This shouldn't cause any problems with WESGR output, since the timing
will still be very close to what it was previously.

With perfetto tracing, this helps with event ordering. The final flush
event will now come after the damage accumulation function calls instead
of before. This is better when we add flow lines in a following commit.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>