backend-drm: do not disable underlay planes based on b->format
Since commit "backend-drm: improve code that chooses output->format", we
changed how b->format (the GBM format from the config file) is used.

There are config options in weston.ini that allow us to ignore the GBM
format set. So what matters to us is which output->format is being used,
and we compute that in drm_output_pick_format_egl() and
drm_output_pick_format_pixman().

So remove these checks that disable underlay planes based on b->format,
keeping only the ones that depend on the output->format selection.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>