gl-renderer: move color mapping asserts to load config
The frequency of checking remaing the same, but we can use an already
needed switch statement and do not need one just for the asserts. This
makes gl_shader_config_set_color_transform() nice and short.

The MAPPING_MATRIX assert was useless, the variable is set right above.

Asserting the mapping type is valid happens already in
gl_shader_load_config_mapping().

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>